Litt mer forklarende oppsett av resttemplate interceptors og servlet filter
Vi diskutrete i teamet og snakket om at vi gjerne skulle kommet oss bort fra spring automagi.
Ser ut som det ikke er så lett når det gjelder servlet filterne. Man kan wire de på flere måter, men ingen av disse virket veldig mye mer eksplisitt.
Når det gjelder resttemplate så er denne endringen et steg på veien.
Vi kan som neste steg koble fra RestTemplateCustomizer beans og heller sette disse eksplisitt alle steder vi bygger en resttemplate
